requirements.txt
numpy==1.26.4
torch==2.6.0
hydra-core==1.3.2
omegaconf==2.3.0
gymnasium==1.2.3
tqdm==4.67.3
wandb==0.25.1
matplotlib==3.10.8
tensorboard==2.20.0
